syndication code cleanup and refactoring.
