+++ server/liveblog/syndication/utils.py
@@ -3,14 +3,17 @@ import uuid
 import hmac
 import logging
 import requests
-from eve.io.mongo import MongoJSONEncoder
+import tempfile
 from bson import ObjectId
 from hashlib import sha1
 from flask import make_response
-import tempfile
+from eve.io.mongo import MongoJSONEncoder
 from .exceptions import APIConnectionError, DownloadError
 from requests.exceptions import RequestException
 from requests.packages.urllib3.exceptions import MaxRetryError
+from superdesk import get_resource_service
+from superdesk.metadata.item import ITEM_TYPE, CONTENT_TYPE
+from .tasks import fetch_image
 
 
 logger = logging.getLogger('superdesk')
@@ -96,3 +99,137 @@ def fetch_url(url, timeout=5):
             fd.write(chunk)
     fd.seek(0)
     return fd
+
+
+def extract_post_items(original_doc):
+    items_service = get_resource_service('items')
+    item_type = original_doc.get(ITEM_TYPE, '')
+    if item_type != CONTENT_TYPE.COMPOSITE:
+        raise NotImplementedError('Post item_type "{}" not supported.'.format(item_type))
+
+    items = []
+    for group in original_doc['groups']:
+        if group['id'] == 'main':
+            for ref in group['refs']:
+                item = items_service.find_one(req=None, guid=ref['guid'])
+                item_type = item['item_type']
+                data = {
+                    'text': item['text'],
+                    'item_type': item_type,
+                    'meta': item.get('meta', {})
+                }
+                items.append(data)
+    return items
+
+
+def get_html_from_image_data(data, meta):
+    srcset = []
+    for value in data['renditions'].values():
+        srcset.append('{} {}w'.format(value['href'], value['width']))
+
+    caption = meta['caption']
+    full_caption = caption
+    if meta['credit']:
+        full_caption = '{} Credit: {}'.format(caption, meta['credit'])
+
+    return ''.join([
+        '<figure>',
+        '   <img src="{}" alt="{}" srcset="{}" />'.format(
+            data['renditions']['viewImage']['href'],
+            meta['caption'],
+            ','.join(srcset)
+        ),
+        '   <figcaption>{}</figcaption>'.format(full_caption),
+        '</figure>'
+    ])
+
+
+def fetch_and_create_image_item(renditions, **meta):
+    try:
+        image_url = renditions['original']['href']
+        mimetype = renditions['original']['mimetype']
+    except KeyError:
+        raise DownloadError('Unable to get original image from renditions: {}'.format(renditions))
+
+    item_data = dict()
+    item_data['type'] = 'picture'
+    item_data['media'] = fetch_image(image_url, mimetype)
+    archive_service = get_resource_service('archive')
+    item_id = archive_service.post([item_data])[0]
+    logger.warning('Image posted as archive: "{}"'.format(item_id))
+    archive = archive_service.find_one(req=None, _id=item_id)
+    return {
+        'item_type': 'image',
+        'meta': {
+            'media': {
+                '_id': item_id,
+                'renditions': archive['renditions']
+            },
+            'caption': meta.get('caption', ''),
+            'credit': meta.get('credit', '')
+        },
+        'text': get_html_from_image_data(archive, meta)
+    }
+
+
+def _create_producer_post_id(in_syndication, post_id):
+    """Helps to denormalize syndication producer blog post data and provide unique value for producer_post_id field."""
+    return '{}:{}:{}'.format(
+        in_syndication['producer_id'],
+        in_syndication['producer_blog_id'],
+        post_id
+    )
+
+
+def create_syndicated_blog_post(producer_post, items, in_syndication, post_status=None):
+    post_items = []
+    for item in items:
+        meta = item.pop('meta')
+        if item['item_type'] == 'image':
+            item = fetch_and_create_image_item(
+                renditions=meta['media']['renditions'],
+                caption=meta['caption'],
+                credit=meta['credit']
+            )
+        item['blog'] = in_syndication['blog_id']
+        post_items.append(item)
+
+    items_service = get_resource_service('blog_items')
+    item_ids = items_service.post(post_items)
+    item_refs = []
+    for i, item_id in enumerate(item_ids):
+        item_refs.append({
+            'residRef': str(item_id)
+        })
+
+    if not post_status:
+        auto_publish = in_syndication.get('auto_publish', False)
+        if auto_publish:
+            post_status = 'open'
+        else:
+            post_status = 'submitted'
+
+    new_post = {
+        'blog': in_syndication['blog_id'],
+        'groups': [
+            {
+                'id': 'root',
+                'role': 'grpRole:NEP',
+                'refs': [
+                    {'idRef': 'main'}
+                ]
+            },
+            {
+                'id': 'main',
+                'role': 'grpRole:Main',
+                'refs': item_refs
+            }
+        ],
+        'highlight': False,
+        'particular_type': 'post',
+        'producer_post_id': _create_producer_post_id(in_syndication, producer_post['_id']),
+        'sticky': False,
+        'syndication_in': in_syndication['_id'],
+        'post_status': post_status
+    }
+    return new_post
